Fix jsdoc generated error types

+++ scripts/run-integration-tests.sh
@@ -5,4 +5,8 @@ set -euo pipefail
 ./build/downloads/bin/conjure-verification-server ./build/resources/verification-server-test-cases.json ./build/resources/verification-server-api.conjure.json &
 SERVER_PID=$!
 yarn karma start --single-run --browsers ChromeHeadless karma.conf.js
-kill -kill ${SERVER_PID}
+
+if ps -p ${SERVER_PID} > /dev/null
+then
+    kill -kill ${SERVER_PID}
+fi
